requirements.txt
pinboard==2.1.8
PyYAML==5.3.1
